requirements.txt
git+https://github.com/domdfcoding/cp2077-extractor
click>=8.3.1
consolekit>=1.9.0
dom-toml>=2.1.0
domdf-python-tools>=3.10.0
pillow>=11.0.0
